컴포넌트 통합
