--- docs/doc.md
@@ -1,7 +1,0 @@
-# DB
-
-```ts
-interface {
-    
-}
-```
