--- docs/test.md
@@ -1,9 +1,0 @@
-|in A| in B| in C| out D|
-|:-:|:-:|:-:|:-:|
-|0|0|0|0|
-|0|0|1|0|
-|0|1|0|0|
-|1|0|0|0|
-|1|0|1|0|
-|1|1|0|0|
-|1|1|1|1
