+++ src/routes/(main)/doc/log/[id]/+page.server.ts
@@ -1,0 +1,3 @@
+export async function load({params}){
+    
+}
