+++ src/routes/(main)/notice/+page.svelte
@@ -1,18 +1,35 @@
 <script lang="ts">
+    import { goto } from "$app/navigation";
     import { page } from "$app/stores";
+    import PageSelector from "$lib/components/common/PageSelector.svelte";
     import PageTitle from "$lib/components/common/PageTitle.svelte";
     import NoticeList from "$lib/components/page/notice/NoticeList.svelte";
-    import NoticePageSelector from "$lib/components/page/notice/NoticePageSelector.svelte";
 
-    let {data} = $props();
+    let { data } = $props();
 
     const pageNum = Number($page.url.searchParams.get("page")) || 1;
+
+    function movePage(p: number){
+        const searchParam = new URLSearchParams(location.search);
+
+        searchParam.set("page", p.toString());
+
+        goto(`/notice?${searchParam}`);
+
+        window.scrollTo({
+            top: 0,
+        });
+    }
 </script>
 
-<PageTitle title="공지"/>
+<PageTitle title="공지" />
 <h1>공지</h1>
-<NoticeList notices={data.notices}/>
-<NoticePageSelector {pageNum} length={data.count} />
+<NoticeList notices={data.notices} />
+<PageSelector
+    {pageNum}
+    length={data.count}
+    {movePage}
+/>
 
 <style>
     h1 {
