+++ src/routes/(main)/song/+page.svelte
@@ -2,13 +2,14 @@
     import { type SongLang } from "$lib/module/common/song/types";
     import SongLanguageSelector from "$lib/components/page/song/SongLanguageSelector.svelte";
     import SongList from "$lib/components/page/song/SongList.svelte";
-    import PageSelector from "$lib/components/page/song/PageSelector.svelte";
     import AddSongButton from "$lib/components/page/song/AddSongButton.svelte";
     import SearchBox from "$lib/components/page/song/SearchBox.svelte";
     import { navigating } from "$app/stores";
     import Loading from "$lib/components/common/Loading.svelte";
     import PageTitle from "$lib/components/common/PageTitle.svelte";
     import { getI18N, getLang } from "$lib/module/common/i18n/i18n.js";
+    import PageSelector from "$lib/components/common/PageSelector.svelte";
+    import { goto } from "$app/navigation";
 
     let {data} = $props();
 
@@ -21,6 +22,18 @@
 
     const lang = getLang();
     let titleI18n = $derived(getI18N($lang).title["/song"]);
+
+    function movePage(p: number) {
+        const searchParam = new URLSearchParams(location.search);
+
+        searchParam.set("page", p.toString());
+
+        goto(`/song?${searchParam}`);
+
+        window.scrollTo({
+            top: 0,
+        });
+    }
 </script>
 
 <PageTitle title={titleI18n} />
@@ -33,6 +46,6 @@
         <SongLanguageSelector bind:songLang />
     {/if}
     <SongList {songLang} {songs} />
-    <PageSelector {pageNum} {length} />
+    <PageSelector {pageNum} {length} {movePage}/>
     <AddSongButton />
 {/if}
